Refactor code structure
Refactored the codebase for improved organization and readability. Consider reviewing components for further optimization as some files have grown lengthy.
[skip gpt_engineer]
